+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed incorrect minisign command on identity proof page.
+- Fixed invalid tooltip for searched unicode emoji.
 
 ## [4.11.0] - 2025-10-09